+++ src/utils.ts
@@ -8,6 +8,11 @@ export function formatMessage(obj: unknown, message = '>>>'): string {
   return `${message} >> ${JSON.stringify(obj, null, 2)}`
 }
 
+export function randomlyReturnReviewer(reviewers: string[]): string {
+  const randomIndex = Math.floor(Math.random() * reviewers.length)
+  return reviewers[randomIndex]
+}
+
 function filterAuthorFromReviewerList(
   list: string[],
   author: string
@@ -39,4 +44,29 @@ export async function listEligibleReviewers(): Promise<string[]> {
 
   const collaborators = data.map(({ login }) => login)
   return filterAuthorFromReviewerList(collaborators, author)
+}
+
+export async function requestReviewer(): Promise<unknown> {
+  // Setup
+  const { pull_request, repository }: WebhookPayload = github.context.payload
+  const pull_number: number = pull_request?.number || 0
+  const owner: string = repository?.owner.login || ''
+  const repo: string = repository?.name || ''
+
+  const github_token: string = core.getInput('github_token')
+  const octokit: InstanceType<typeof GitHub> = github.getOctokit(github_token)
+
+  // Selects reviewer
+  const reviewers = await listEligibleReviewers()
+  const selectedReviewer = randomlyReturnReviewer(reviewers)
+
+  // Request review
+  const result = await octokit.pulls.requestReviewers({
+    reviewers: [selectedReviewer],
+    pull_number,
+    owner,
+    repo
+  })
+
+  return result
 }
